Change DictSpace to an ordered dict
Ordered the elements of a DictSpace. This is needed because if using a
map, the keys are unordered. In such a case, consecutive calls to
High(), Low(), Sample() would give different results, and there would be
no way of knowing which elements in the returned sliced corresponeded to
which spaces that the DictSpace contained.

Now, if the DictSpace contains a number of spaces, it holds them in
the order of insertion. This way, the methods will always return the
same values in the same order, and the first *mat.VecDense in the
returned slice from High()/Low()/Sample() will always refer to the first
space in the DictSpace. The second returned slice to the second space,
and so on.